Fix pbdb_ref_occurrences test
It seems that the "year" parameter in the occs/refs endpoint from
version 1.1 of the API is called "ref_pubyr" in version 1.2, which
also allows to specify a range of years according to the
documentation.  However, using that parameter produces a 500 Internal
Server Error response at the time of this commit.  Since the test is
not about that, I changed the parameter for a different one that
returns a few records.