Make .extract_temporal_range() work with compact field names
The reason it was not working properly with compact field names (which
are returned by the API by default) was that, in compact form, the
"rnk" column ("accepted_rank") is returned as a rank code instead of
the name of the rank as a string.

See <https://paleobiodb.org/data1.2/config.txt?show=ranks> for the
corresponding rank codes.