Run `.pbdb_setup()` on load
This makes the Collate field in DESCRIPTION no longer needed, and
populates the `.private_pk_cache` environment as intended without
invoking the function in the `pbdb_rest_api_setup.R` file.

However, all this code involving a cache object might need some
rewriting, because it is not very useful in the current state of the
package.