Simplify package architecture
All endpoints were using the same URI builder function, so there is no
need to maintain a data structure with specific functions and
mandatory parameters for each endpoint that is constructed at load
time.

Mandatory parameters could be checked in the exported functions that
act as interfaces for each endpoint, and in any case the API itself
will reply with a (hopefully) descriptive error message if some
mandatory parameter is not specified.