Replace rjson dependency by jsonlite
There are a couple of reasons for this change.  First, rjson depends
on a recent version of R (>= 4.4.0), which is the current released
minor version, while jsonlite should be able to work with older
versions.  Second, as of the time of this writing, the CRAN task view
for Web Technologies and Services recommends using the jsonlite
package.

https://cran.r-project.org/web/views/WebTechnologies.html

Also, fromJSON() in jsonlite coerces the records in the response to a
data.frame by default, which is in principle quite convenient (and
gtools::smartbind() does not seem to be needed anymore), although it
is possible that in some cases this gives unexpected results.  Some
new bugs could appear because of this.